learn_pytest_bdd/requirements.txt
pytest
pytest-bdd
pytest-selenium
